refactor: handleDeclaration, handleDelete 생성
