+++ src/pages/PostDetail/index.jsx
@@ -6,7 +6,7 @@ import CommentBar from "../../components/CommentBar";
 import Comment from "./Comment";
 import UserPostDetail from "./UserPostDetail";
 import { useParams } from "react-router-dom";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useGetData } from "../../hooks/useGetData";
 
 export default function PostDetail() {
@@ -15,19 +15,17 @@ export default function PostDetail() {
   const { data: commentData, getData: commentGetData } = useGetData();
   const postUrl = `${process.env.REACT_APP_API_KEY}/post/${postkey}`;
   const commentUrl = `${process.env.REACT_APP_API_KEY}/post/${postkey}/comments`;
-
-  useEffect(() => {
-    postGetData(postUrl);
-  }, []);
+  const [trigger, setTrigger] = useState(false);
 
   const setCommentList = async () => {
     const res = await commentGetData(commentUrl);
   };
 
   useEffect(() => {
+    postGetData(postUrl);
     commentGetData(commentUrl);
     console.log(commentData);
-  }, []);
+  }, [trigger]);
 
   return (
     <>
@@ -42,7 +40,7 @@ export default function PostDetail() {
             <Comment
               data={mapData}
               commentId={mapData.id}
-              setCommentList={setCommentList}
+              setTrigger={setTrigger}
             />
           ))}
       </MainContainer>
