+++ src/utils/handleDelete.js
@@ -1,0 +1,22 @@
+import axios from "axios";
+
+const userInfo = JSON.parse(localStorage.getItem("userinfo"));
+
+export const handleDelete = async (e, handlCloseClick, setTrigger, url) => {
+  e.preventDefault();
+  try {
+    const res = await axios.delete(url, {
+      headers: {
+        Authorization: `Bearer ${userInfo.token}`,
+        "Content-type": "application/json",
+      },
+    });
+    if (res.status === 200) {
+      setTrigger((prev) => !prev);
+      handlCloseClick();
+      alert("삭제되었습니다.");
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
